storage: move externally visible source pipeline types to types.rs
This is pure code movement and fixing up use statements.
