+++ src/storage/src/source/types.rs
@@ -1,0 +1,516 @@
+// Copyright Materialize, Inc. and contributors. All rights reserved.
+//
+// Use of this software is governed by the Business Source License
+// included in the LICENSE file.
+//
+// As of the Change Date specified in that file, in accordance with
+// the Business Source License, use of this software will be governed
+// by the Apache License, Version 2.0.
+
+//! Types related to the source ingestion pipeline/framework.
+
+// https://github.com/tokio-rs/prost/issues/237
+// #![allow(missing_docs)]
+
+use std::collections::HashMap;
+use std::fmt;
+use std::fmt::Debug;
+use std::rc::Rc;
+use std::time::Duration;
+
+use async_trait::async_trait;
+use differential_dataflow::Hashable;
+use futures::stream::LocalBoxStream;
+use prometheus::core::{AtomicI64, AtomicU64};
+use serde::{Deserialize, Serialize};
+use timely::dataflow::channels::pact::{Exchange, ParallelizationContract};
+use timely::scheduling::activate::SyncActivator;
+use timely::scheduling::ActivateOnDrop;
+use timely::Data;
+
+use mz_avro::types::Value;
+use mz_expr::PartitionId;
+use mz_ore::metrics::{CounterVecExt, DeleteOnDropCounter, DeleteOnDropGauge, GaugeVecExt};
+use mz_repr::{Diff, GlobalId, Row, Timestamp};
+
+use crate::source::healthcheck::SourceStatusUpdate;
+use crate::source::metrics::SourceBaseMetrics;
+use crate::types::connections::ConnectionContext;
+use crate::types::errors::{DecodeError, SourceErrorDetails};
+use crate::types::sources::encoding::SourceDataEncoding;
+use crate::types::sources::{MzOffset, SourceConnection};
+
+/// This trait defines the interface between Materialize and external sources,
+/// and must be implemented for every new kind of source.
+///
+/// ## Contract between [`SourceReader`] and the ingestion framework
+///
+/// A source reader uses updates emitted from
+/// [`SourceReader::next`]/[`SourceReader::get_next_message`] to update the
+/// ingestion framework about new updates retrieved from the external system and
+/// about its internal state.
+///
+/// The framework will spawn a [`SourceReader`] on each timely worker. It is the
+/// responsibility of the reader to figure out which of the partitions (if any)
+/// it is responsible for reading using [`crate::source::responsible_for`].
+///
+/// The reader implicitly is given a capability for emitting updates for each
+/// new partition (identified by a [`PartitionId`]) that it discovers. It must
+/// downgrade those capabilities by either emitting updates for those partitions
+/// that it is responsible for or by emitting a
+/// [`SourceMessageType::DropPartitionCapabilities`] for those partitions which
+/// it is not responsible for.
+//
+// TODO: this trait is still a little too Kafka-centric, specifically the concept of
+// a "partition" is baked into this trait and introduces some cognitive overhead as
+// we are forced to treat things like file sources as "single-partition"
+#[async_trait(?Send)]
+pub trait SourceReader {
+    type Key: timely::Data + MaybeLength;
+    type Value: timely::Data + MaybeLength;
+    type Diff: timely::Data;
+
+    /// Create a new source reader.
+    ///
+    /// This function returns the source reader and optionally, any "partition" it's
+    /// already reading. In practice, the partition is only non-None for static sources
+    /// that either don't truly have partitions or have a fixed number of partitions.
+    fn new(
+        source_name: String,
+        source_id: GlobalId,
+        worker_id: usize,
+        worker_count: usize,
+        consumer_activator: SyncActivator,
+        connection: SourceConnection,
+        restored_offsets: Vec<(PartitionId, Option<MzOffset>)>,
+        encoding: SourceDataEncoding,
+        metrics: crate::source::metrics::SourceBaseMetrics,
+        connection_context: ConnectionContext,
+    ) -> Result<Self, anyhow::Error>
+    where
+        Self: Sized;
+
+    /// Returns the next message available from the source.
+    ///
+    /// Note that implementers are required to present messages in strictly ascending offset order
+    /// within each partition.
+    async fn next(
+        &mut self,
+        timestamp_granularity: Duration,
+    ) -> Option<Result<SourceMessageType<Self::Key, Self::Value, Self::Diff>, SourceReaderError>>
+    {
+        // Compatiblity implementation that delegates to the deprecated [Self::get_next_method]
+        // call. Once all source implementations have been transitioned to implement
+        // [SourceReader::next] directly this provided implementation should be removed and the
+        // method should become a required method.
+        loop {
+            match self.get_next_message() {
+                Ok(NextMessage::Ready(msg)) => return Some(Ok(msg)),
+                Err(err) => return Some(Err(err)),
+                // There was a temporary hiccup in getting messages, check again asap.
+                Ok(NextMessage::TransientDelay) => {
+                    tokio::time::sleep(Duration::from_millis(1)).await
+                }
+                // There were no new messages, check again after a delay
+                Ok(NextMessage::Pending) => tokio::time::sleep(timestamp_granularity).await,
+                Ok(NextMessage::Finished) => return None,
+            }
+        }
+    }
+
+    /// Returns the next message available from the source.
+    ///
+    /// Note that implementers are required to present messages in strictly ascending offset order
+    /// within each partition.
+    ///
+    /// # Deprecated
+    ///
+    /// Source implementation should implement the async [SourceReader::next] method instead.
+    fn get_next_message(
+        &mut self,
+    ) -> Result<NextMessage<Self::Key, Self::Value, Self::Diff>, SourceReaderError> {
+        Ok(NextMessage::Pending)
+    }
+
+    /// Returns an adapter that treats the source as a stream.
+    ///
+    /// The stream produces the messages that would be produced by repeated calls to `next`.
+    fn into_stream<'a>(
+        mut self,
+        timestamp_granularity: Duration,
+    ) -> LocalBoxStream<
+        'a,
+        Result<SourceMessageType<Self::Key, Self::Value, Self::Diff>, SourceReaderError>,
+    >
+    where
+        Self: Sized + 'a,
+    {
+        Box::pin(async_stream::stream!({
+            while let Some(msg) = self.next(timestamp_granularity).await {
+                yield msg;
+            }
+        }))
+    }
+}
+
+/// A `SourceToken` manages interest in a source.
+///
+/// When the `SourceToken` is dropped the associated source will be stopped.
+pub struct SourceToken {
+    pub(crate) _activator: Rc<ActivateOnDrop<()>>,
+}
+
+pub enum NextMessage<Key, Value, Diff> {
+    Ready(SourceMessageType<Key, Value, Diff>),
+    Pending,
+    TransientDelay,
+    Finished,
+}
+
+/// A wrapper around [`SourceMessage`] that allows [`SourceReader`]'s to
+/// communicate additional "maintenance" messages.
+pub enum SourceMessageType<Key, Value, Diff> {
+    /// Communicate that this [`SourceMessage`] is the final
+    /// message its its offset.
+    Finalized(SourceMessage<Key, Value, Diff>),
+    /// Communicate that more [`SourceMessage`]'s
+    /// will come later at the same offset as this one.
+    InProgress(SourceMessage<Key, Value, Diff>),
+    /// Information about the source status
+    SourceStatus(SourceStatusUpdate),
+    /// Signals that this [`SourceReader`] instance will never emit
+    /// messages/updates for a given partition anymore. This is similar enough
+    /// to a timely operator dropping a capability, hence the naming.
+    ///
+    /// We need these to compute a "global" source upper, when determining
+    /// completeness of a timestamp.
+    DropPartitionCapabilities(Vec<PartitionId>),
+}
+
+/// Source-agnostic wrapper for messages. Each source must implement a
+/// conversion to Message.
+pub struct SourceMessage<Key, Value, Diff> {
+    /// Partition from which this message originates
+    pub partition: PartitionId,
+    /// Materialize offset of the message (1-indexed)
+    pub offset: MzOffset,
+    /// The time that an external system first observed the message
+    ///
+    /// Milliseconds since the unix epoch
+    pub upstream_time_millis: Option<i64>,
+    /// The message key
+    pub key: Key,
+    /// The message value
+    pub value: Value,
+    /// Headers, if the source is configured to pass them along. If it is, but there are none, it
+    /// passes `Some([])`
+    pub headers: Option<Vec<(String, Option<Vec<u8>>)>>,
+
+    /// Allow sources to optionally output a specific differential
+    /// `diff` value. Defaults to `+1`.
+    ///
+    /// Only supported with `SourceEnvelope::None`
+    pub specific_diff: Diff,
+}
+
+impl fmt::Debug for SourceMessage<(), Option<Vec<u8>>, ()> {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        f.debug_struct("SourceMessage")
+            .field("partition", &self.partition)
+            .field("offset", &self.offset)
+            .field("upstream_time_millis", &self.upstream_time_millis)
+            .finish()
+    }
+}
+
+impl fmt::Debug for SourceMessage<Option<Vec<u8>>, Option<Vec<u8>>, ()> {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        f.debug_struct("SourceMessage")
+            .field("partition", &self.partition)
+            .field("offset", &self.offset)
+            .field("upstream_time_millis", &self.upstream_time_millis)
+            .field("key[present]", &self.key.is_some())
+            .field("value[present]", &self.value.is_some())
+            .finish()
+    }
+}
+
+/// A record produced by a source
+#[derive(Clone, Serialize, Debug, Deserialize)]
+pub struct SourceOutput<K, V, D>
+where
+    K: Data,
+    V: Data,
+{
+    /// The record's key (or some empty/default value for sources without the concept of key)
+    pub key: K,
+    /// The record's value
+    pub value: V,
+    /// The position in the partition described by the `partition` in the source
+    /// (e.g., Kafka offset, file line number, monotonic increasing
+    /// number, etc.)
+    pub position: MzOffset,
+    /// The time the record was created in the upstream system, as milliseconds since the epoch
+    pub upstream_time_millis: Option<i64>,
+    /// The partition of this message, present iff the partition comes from Kafka
+    pub partition: PartitionId,
+    /// Headers, if the source is configured to pass them along. If it is, but there are none, it
+    /// passes `Some([])`
+    pub headers: Option<Vec<(String, Option<Vec<u8>>)>>,
+
+    /// Indicator for what the differential `diff` value
+    /// for this decoded message should be
+    pub diff: D,
+}
+
+impl<K, V, D> SourceOutput<K, V, D>
+where
+    K: Data,
+    V: Data,
+{
+    /// Build a new SourceOutput
+    pub fn new(
+        key: K,
+        value: V,
+        position: MzOffset,
+        upstream_time_millis: Option<i64>,
+        partition: PartitionId,
+        headers: Option<Vec<(String, Option<Vec<u8>>)>>,
+        diff: D,
+    ) -> SourceOutput<K, V, D> {
+        SourceOutput {
+            key,
+            value,
+            position,
+            upstream_time_millis,
+            partition,
+            headers,
+            diff,
+        }
+    }
+}
+
+impl<K, V, D> SourceOutput<K, V, D>
+where
+    K: Data + Serialize + for<'a> Deserialize<'a> + Send + Sync,
+    V: Data + Serialize + for<'a> Deserialize<'a> + Send + Sync,
+    D: Data + Serialize + for<'a> Deserialize<'a> + Send + Sync,
+{
+    /// A parallelization contract that hashes by positions (if available)
+    /// and otherwise falls back to hashing by value. Values can be just as
+    /// skewed as keys, whereas positions are generally known to be unique or
+    /// close to unique in a source. For example, Kafka offsets are unique per-partition.
+    /// Most decode logic should use this instead of `key_contract`.
+    pub fn position_value_contract() -> impl ParallelizationContract<Timestamp, Self>
+    where
+        V: Hashable<Output = u64>,
+    {
+        Exchange::new(|x: &Self| x.position.hashed())
+    }
+}
+
+/// The output of the decoding operator
+#[derive(Debug, PartialEq, Eq, Hash, Clone, Serialize, Deserialize)]
+pub struct DecodeResult {
+    /// The decoded key
+    pub key: Option<Result<Row, DecodeError>>,
+    /// The decoded value, as well as the the
+    /// differential `diff` value for this value, if the value
+    /// is present and not and error.
+    pub value: Option<Result<(Row, Diff), DecodeError>>,
+    /// The index of the decoded value in the stream
+    pub position: MzOffset,
+    /// The time the record was created in the upstream system, as milliseconds since the epoch
+    pub upstream_time_millis: Option<i64>,
+    /// The partition this record came from
+    pub partition: PartitionId,
+    /// If this is a Kafka stream, the appropriate metadata
+    // TODO(bwm): This should probably be statically different for different streams, or we should
+    // propagate whether metadata is requested into the decoder
+    pub metadata: Row,
+}
+
+/// A structured error for `SourceReader::get_next_message`
+/// implementors. Also implements `From<anyhow::Error>`
+/// for convenience.
+pub struct SourceReaderError {
+    pub inner: SourceErrorDetails,
+}
+
+impl From<anyhow::Error> for SourceReaderError {
+    fn from(e: anyhow::Error) -> Self {
+        SourceReaderError {
+            inner: SourceErrorDetails::Other(format!("{}", e)),
+        }
+    }
+}
+
+/// Source-specific Prometheus metrics
+pub struct SourceMetrics {
+    /// Number of times an operator gets scheduled
+    pub(crate) operator_scheduled_counter: DeleteOnDropCounter<'static, AtomicU64, Vec<String>>,
+    /// Value of the capability associated with this source
+    pub(crate) capability: DeleteOnDropGauge<'static, AtomicU64, Vec<String>>,
+    /// Per-partition Prometheus metrics.
+    pub(crate) partition_metrics: HashMap<PartitionId, PartitionMetrics>,
+    source_name: String,
+    source_id: GlobalId,
+    base_metrics: SourceBaseMetrics,
+}
+
+impl SourceMetrics {
+    /// Initialises source metrics for a given (source_id, worker_id)
+    pub fn new(
+        base: &SourceBaseMetrics,
+        source_name: &str,
+        source_id: GlobalId,
+        worker_id: &str,
+    ) -> SourceMetrics {
+        let labels = &[
+            source_name.to_string(),
+            source_id.to_string(),
+            worker_id.to_string(),
+        ];
+        SourceMetrics {
+            operator_scheduled_counter: base
+                .source_specific
+                .operator_scheduled_counter
+                .get_delete_on_drop_counter(labels.to_vec()),
+            capability: base
+                .source_specific
+                .capability
+                .get_delete_on_drop_gauge(labels.to_vec()),
+            partition_metrics: Default::default(),
+            source_name: source_name.to_string(),
+            source_id,
+            base_metrics: base.clone(),
+        }
+    }
+
+    /// Log updates to which offsets / timestamps read up to.
+    pub fn record_partition_offsets(
+        &mut self,
+        offsets: HashMap<PartitionId, (MzOffset, Timestamp, i64)>,
+    ) {
+        for (partition, (offset, timestamp, count)) in offsets {
+            let metric = self
+                .partition_metrics
+                .entry(partition.clone())
+                .or_insert_with(|| {
+                    PartitionMetrics::new(
+                        &self.base_metrics,
+                        &self.source_name,
+                        self.source_id,
+                        &partition,
+                    )
+                });
+
+            metric.messages_ingested.inc_by(count);
+
+            metric.record_offset(
+                &self.source_name,
+                self.source_id,
+                &partition,
+                offset.offset,
+                timestamp as i64,
+            );
+        }
+    }
+}
+
+/// Partition-specific metrics, recorded to both Prometheus and a system table
+pub struct PartitionMetrics {
+    /// Highest offset that has been received by the source and timestamped
+    pub(crate) offset_ingested: DeleteOnDropGauge<'static, AtomicU64, Vec<String>>,
+    /// Highest offset that has been received by the source
+    pub(crate) offset_received: DeleteOnDropGauge<'static, AtomicU64, Vec<String>>,
+    /// Value of the highest timestamp that is closed (for which all messages have been ingested)
+    pub(crate) closed_ts: DeleteOnDropGauge<'static, AtomicU64, Vec<String>>,
+    /// Total number of messages that have been received by the source and timestamped
+    pub(crate) messages_ingested: DeleteOnDropCounter<'static, AtomicI64, Vec<String>>,
+    pub(crate) last_offset: u64,
+    pub(crate) last_timestamp: i64,
+}
+
+impl PartitionMetrics {
+    /// Record the latest offset ingested high-water mark
+    fn record_offset(
+        &mut self,
+        _source_name: &str,
+        _source_id: GlobalId,
+        _partition_id: &PartitionId,
+        offset: u64,
+        timestamp: i64,
+    ) {
+        self.offset_received.set(offset);
+        self.offset_ingested.set(offset);
+        self.last_offset = offset;
+        self.last_timestamp = timestamp;
+    }
+
+    /// Initialises partition metrics for a given (source_id, partition_id)
+    pub fn new(
+        base_metrics: &SourceBaseMetrics,
+        source_name: &str,
+        source_id: GlobalId,
+        partition_id: &PartitionId,
+    ) -> PartitionMetrics {
+        let labels = &[
+            source_name.to_string(),
+            source_id.to_string(),
+            partition_id.to_string(),
+        ];
+        let base = &base_metrics.partition_specific;
+        PartitionMetrics {
+            offset_ingested: base
+                .offset_ingested
+                .get_delete_on_drop_gauge(labels.to_vec()),
+            offset_received: base
+                .offset_received
+                .get_delete_on_drop_gauge(labels.to_vec()),
+            closed_ts: base.closed_ts.get_delete_on_drop_gauge(labels.to_vec()),
+            messages_ingested: base
+                .messages_ingested
+                .get_delete_on_drop_counter(labels.to_vec()),
+            last_offset: 0,
+            last_timestamp: 0,
+        }
+    }
+}
+
+/// Types that implement this trait expose a length function
+pub trait MaybeLength {
+    /// Returns the size of the object
+    fn len(&self) -> Option<usize>;
+}
+
+impl MaybeLength for () {
+    fn len(&self) -> Option<usize> {
+        None
+    }
+}
+
+impl MaybeLength for Vec<u8> {
+    fn len(&self) -> Option<usize> {
+        Some(self.len())
+    }
+}
+
+impl MaybeLength for mz_repr::Row {
+    fn len(&self) -> Option<usize> {
+        Some(self.data().len())
+    }
+}
+
+impl MaybeLength for Value {
+    // Not possible to compute a size in bytes without recursively traversing the entire tree.
+    fn len(&self) -> Option<usize> {
+        None
+    }
+}
+
+impl<T: MaybeLength> MaybeLength for Option<T> {
+    fn len(&self) -> Option<usize> {
+        self.as_ref().and_then(|v| v.len())
+    }
+}
